requirements.txt
pandas
numpy
requests
python-dotenv
streamlit
plotly
